+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Add tablebase support and follow tablebase lines (6.79 +- 4.59 Elo (5-man vs none))
 
+### Misc
+
+* Allow sending only 'go wtime' or 'go btime'
+
 ## [4.0]
 
 * Add late move reductions (100.80 +- 22.10 Elo)